1. 优化了响应请求代码 2. 新增了二级评论功能 3. 对结构的优化

+++ 项目/Blog/server/controller/comment.ts
@@ -4,8 +4,9 @@ import CommentController from "../@types/controller/comment";
 /* 引入comment模型 */
 import commentModel from "../model/comment";
 
-/* 引入jwt工具 */
+/* 引入工具 */
 import Jwt from "../utils/jwt";
+import handelResponse from "../utils/handelResponse";
 
 const commentController: CommentController = {
   // 创建评论
@@ -13,22 +14,25 @@ const commentController: CommentController = {
     try {
       // 获取请求体和answerId
       const info = req.body;
-      const { answerId } = req.params;
-
+      const { answerId, id: commentId } = req.params;
       // 评论创建者id
       const token = req.headers.token as string;
       const { value } = Jwt.verify(token);
-
       // 将commentator整合进info中
       info.commentator = value;
       info.answerId = answerId;
 
-      await commentModel.create(info);
-      res.status(200).json({
-        code: 200,
-        message: "评论创建成功",
-        data: info,
-      });
+      // 如果不是创建二级评论
+      if (!commentId) {
+        const result = await commentModel.create(info);
+        handelResponse(res, result, info);
+      } else {
+        info.replyTo = commentId;
+        const result = await commentModel.findByIdAndUpdate(commentId, {
+          $push: { secondaryComment: info },
+        });
+        handelResponse(res, result, info);
+      }
     } catch (err) {
       next(err);
     }
@@ -36,21 +40,18 @@ const commentController: CommentController = {
   // 修改评论
   updateComment: async (req, res, next) => {
     try {
-      const id = req.params.id;
-      const answerInfo = req.body;
-      const answer = await commentModel.findByIdAndUpdate(id, answerInfo);
-      if (answer) {
-        res.status(200).json({
-          code: 200,
-          message: "评论修改成功",
-          data: answerInfo,
-        });
+      const { id, sId } = req.params;
+      const info = req.body;
+      // 如果不是修改二级评论
+      if (!sId) {
+        const answer = await commentModel.findByIdAndUpdate(id, info);
+        handelResponse(res, answer, info);
       } else {
-        res.status(400).json({
-          code: 400,
-          message: "评论修改失败",
-          data: answer,
-        });
+        const sAnswer = await commentModel.findOneAndUpdate(
+          { secondaryComment: { $elemMatch: { _id: sId } } },
+          { $set: { "secondaryComment.$.content": info.content } }
+        );
+        handelResponse(res, sAnswer, info);
       }
     } catch (err) {
       next(err);
@@ -65,33 +66,33 @@ const commentController: CommentController = {
       // 限制每页有多少条数据
       limit = Math.max((limit as any) * 1, 0);
 
-      // 获取questionId answerId
+      // 获取answerId commentId（为二级评论服务）
       const { answerId } = req.params;
       const commentList = await commentModel
-        // 实现模糊搜索，忽略大小写
         .find({ answerId })
-        // .limit(limit)
-        // .skip(page * limit)
-        .populate("commentator");
-      if (commentList) {
-        res.status(200).json({
-          code: 200,
-          message: "查询评论列表成功",
-          data: commentList,
+        .limit(limit)
+        .skip(page * limit)
+        .populate("commentator")
+        .populate({
+          path: "secondaryComment",
+          populate: { path: "commentator" },
+          perDocumentLimit: 3,
+          // options: { limit: 3 },
         });
-      } else {
-        res.status(400).json({
-          code: 400,
-          message: "查询评论失败",
-        });
-      }
+      handelResponse(res, commentList);
     } catch (err) {
       next(err);
     }
   },
-  // 查询指定评论
+  // 查询指定评论下的二级评论列表
   getComment: async (req, res, next) => {
     try {
+      // 获取页数page
+      let { page = 0, limit = 10 } = req.query;
+      page = Math.max((page as any) * 1, 1) - 1;
+      // 限制每页有多少条数据
+      limit = Math.max((limit as any) * 1, 0);
+
       const id = req.params.id;
       // 获取用户详细信息时，使用?detail=xxx的形式
       let detail = req.query.detail as string;
@@ -101,25 +102,13 @@ const commentController: CommentController = {
           .map((item) => " +" + item)
           .join("");
       }
-      console.log(detail);
-
       const comment = await commentModel
         .findById(id)
+        .limit(limit)
+        .skip(page * limit)
         .select(detail)
         .populate("commentator");
-      if (comment) {
-        res.status(200).json({
-          code: 200,
-          message: "查询指定评论成功",
-          data: comment,
-        });
-      } else {
-        res.status(400).json({
-          code: 400,
-          message: "查询指定评论失败",
-          data: { id },
-        });
-      }
+      handelResponse(res, comment);
     } catch (err) {
       next(err);
     }
@@ -127,19 +116,17 @@ const commentController: CommentController = {
   // 删除指定评论
   deleteComment: async (req, res, next) => {
     try {
-      const id = req.params.id;
-      const data = await commentModel.findByIdAndDelete(id);
-      if (data) {
-        res.status(200).json({
-          code: 200,
-          message: "删除评论成功",
-        });
+      const { id, sId } = req.params;
+      // 如果不是二级评论
+      if (!sId) {
+        const data = await commentModel.findByIdAndDelete(id);
+        handelResponse(res, data);
       } else {
-        res.status(400).json({
-          code: 400,
-          message: "删除评论失败",
-          data,
-        });
+        const sAnswer = await commentModel.findOneAndUpdate(
+          { _id: id },
+          { $pull: { secondaryComment: { _id: sId } } }
+        );
+        handelResponse(res, sAnswer);
       }
     } catch (err) {
       next(err);
